extracted "decisions import" to own file; drop Origin from struct
